Fix playback & leaderboard issues

+++ Script/Singleton/BeatmapListener.cs
@@ -20,7 +20,7 @@ public class BeatmapLegend
 	public int Beatidset { get; set; } = -1; // Beatmap ID for Qlute-made beatmaps
 	public float Bpm { get; set; } = 0.0f;
 	public List<DanceCounter> Dance { get; set; } = new List<DanceCounter>();
-	public int Timetotal { get; set; } = 0; // in milliseconds
+	public float Timetotal { get; set; } = 0; // in milliseconds
 	public float Levelrating { get; set; } = 0.0f; // Level rating for the beatmap
 	public double Accuracy { get; set; } = 0.0; // Accuracy Level for the beatmap (not been used yet)
 	public string Background { get; set; } = null; // Background image path
